+++ packages/types/src/rpc.ts
@@ -13,6 +13,7 @@ export enum RpcMethod {
 
   /* EVM */
   ETH_SEND_TRANSACTION = 'eth_sendTransaction',
+  ETH_SEND_TRANSACTION_BATCH = 'eth_sendTransactionBatch',
   SIGN_TYPED_DATA_V3 = 'eth_signTypedData_v3',
   SIGN_TYPED_DATA_V4 = 'eth_signTypedData_v4',
   SIGN_TYPED_DATA_V1 = 'eth_signTypedData_v1',
@@ -227,11 +228,7 @@ export type SigningData =
       account: string;
       data: BitcoingSignTxData;
     }
-  | {
-      type: RpcMethod.ETH_SEND_TRANSACTION;
-      account: string;
-      data: TransactionRequest;
-    }
+  | SigningData_EthSendTx
   | {
       type: RpcMethod.ETH_SIGN | RpcMethod.PERSONAL_SIGN;
       account: string;
@@ -275,22 +272,49 @@ export type SigningData =
       };
     };
 
+export type SigningData_EthSendTx = {
+  type: RpcMethod.ETH_SEND_TRANSACTION;
+  account: string;
+  data: TransactionRequest;
+};
+
+export type EvmTxBatchUpdateFn = (
+  data: { maxFeeRate?: bigint; maxTipRate?: bigint },
+  txIndex: number,
+) => {
+  displayData: DisplayData;
+  signingRequests: {
+    displayData: DisplayData;
+    signingData: SigningData_EthSendTx;
+  }[];
+};
+
 export type EvmTxUpdateFn = (data: {
   maxFeeRate?: bigint;
   maxTipRate?: bigint;
   approvalLimit?: Hex; // as hexadecimal, 0x-prefixed
-}) => { displayData: DisplayData; signingData: Extract<SigningData, { type: RpcMethod.ETH_SEND_TRANSACTION }> };
+}) => { displayData: DisplayData; signingData: SigningData_EthSendTx };
 
 export type BtcTxUpdateFn = (data: { feeRate?: number }) => {
   displayData: DisplayData;
   signingData: Extract<SigningData, { type: RpcMethod.BITCOIN_SEND_TRANSACTION }>;
 };
 
+export type SigningRequest<Data = SigningData> = {
+  displayData: DisplayData;
+  signingData: Data;
+  updateTx?: EvmTxUpdateFn | BtcTxUpdateFn;
+};
+
 export type ApprovalParams = {
   request: RpcRequest;
+} & SigningRequest;
+
+export type BatchApprovalParams = {
+  request: RpcRequest;
+  signingRequests: SigningRequest<SigningData_EthSendTx>[];
   displayData: DisplayData;
-  signingData: SigningData;
-  updateTx?: EvmTxUpdateFn | BtcTxUpdateFn;
+  updateTx: EvmTxBatchUpdateFn;
 };
 
 /**
@@ -306,7 +330,10 @@ export type ApprovalParams = {
  * "eth_sendTransaction" calls, which means that all we'll get in response
  * from them is the transaction hash (not a signed tx).
  */
-export type SigningResult = { signedData: string } | { txHash: string };
+type SignedData = { signedData: string };
+type BroadcastedTx = { txHash: string };
+
+export type SigningResult = SignedData | BroadcastedTx;
 
 export type ApprovalResponse =
   | {
@@ -314,8 +341,17 @@ export type ApprovalResponse =
     }
   | SigningResult;
 
+export type BatchApprovalResponse =
+  | { error: RpcError }
+  | {
+      result: SignedData[];
+    };
 export interface ApprovalController {
   requestApproval: (params: ApprovalParams) => Promise<ApprovalResponse>;
   onTransactionConfirmed: (txHash: Hex, requestId: string) => void;
   onTransactionReverted: (txHash: Hex, requestId: string) => void;
+}
+
+export interface BatchApprovalController extends ApprovalController {
+  requestBatchApproval: (params: BatchApprovalParams) => Promise<BatchApprovalResponse>;
 }