+++ packages-internal/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @internal/utils
 
+## 0.3.5
+
+### Patch Changes
+
+- af15f8d: feat: batch signing for EVM, stable release for all
+
+## 0.3.4
+
+### Patch Changes
+
+- 20e096d: chore: update SDKs
+
 ## 0.3.3
 
 ### Patch Changes