+++ packages/hvm-module/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @avalabs/hvm-module
 
+## 1.1.0
+
+### Patch Changes
+
+- af15f8d: feat: batch signing for EVM, stable release for all
+- Updated dependencies [af15f8d]
+- Updated dependencies [7bee314]
+- Updated dependencies [af15f8d]
+- Updated dependencies [af15f8d]
+  - @avalabs/vm-module-types@1.1.0
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [20e096d]
+  - @avalabs/vm-module-types@0.12.1
+
 ## 1.0.0
 
 ### Major Changes